feat: create hooks for components
