+++ src/helpers/getStep.tsx
@@ -3,9 +3,9 @@ import { UploadStep } from '@/components/UploadStep/index';
 
 export function getStepContent(step: number, handleNext: () => void, onBack: () => void) {
   switch (step) {
-    case 0:
-      return <UploadStep onContinue={handleNext} />;
     case 1:
+      return <UploadStep onContinue={handleNext} />;
+    case 0:
       return <ReviewCustomizeStep onContinue={handleNext} onBack={onBack} />;
 
     default:
