+++ src/i18n/LoginLocales/en/translation.json
@@ -93,8 +93,8 @@
     "guidelines-description": "Physical activity recommendations",
     "recommendations-title": "Supplement recommendations",
     "recommendations-description": "Vitamins and minerals guide",
-    "insights-title": "Medication insights",
-    "insights-description": "General info on relevant medications",
+    "insights-title": "Medication guidance",
+    "insights-description": "Advice on your next medical steps",
     "additional-questions-title": "Additional questions",
     "additional-questions-description": "Ask specific questions about your results or health concerns",
     "your-questions": "Your questions",
